Move parsing functions to utils
